pyproject.toml
[project]
name = "uv-autoeva-reserach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "nbclient>=0.10.2",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
]
